+++ lib/src/dartdap/client/ldap_transformer.dart
@@ -8,6 +8,11 @@ StreamTransformer<Uint8List, LDAPMessage> _createLdapTransformer() {
 
   return new StreamTransformer.fromHandlers(
       handleData: (Uint8List data, EventSink<LDAPMessage> sink) {
+    if (data == null) {
+      loggerRecvBytes.fine("Bytes received: zero");
+      return;
+    }
+
     // Set buf to the bytes to attempt to process: leftover bytes from an
     // earlier data event (if any) plus the new bytes in data.
 
@@ -18,12 +23,12 @@ StreamTransformer<Uint8List, LDAPMessage> _createLdapTransformer() {
       loggerRecvBytes.fine("Bytes received: ${data.length}");
     } else {
       // There were left over bytes: leftover bytes + new data
+      loggerRecvBytes.fine(
+          "Bytes received: ${data.length} (leftover: ${leftover.length})");
       buf = new Uint8List(leftover.length + data.length);
       buf.setRange(0, leftover.length, leftover);
       buf.setRange(leftover.length, buf.length, data);
       leftover = null;
-      loggerRecvBytes.fine(
-          "Bytes received: ${data.length} (+${leftover.length} leftover)");
     }
 
     if (Level.FINEST >= loggerRecvBytes.level) {
@@ -113,7 +118,8 @@ StreamTransformer<Uint8List, LDAPMessage> _createLdapTransformer() {
         leftover = buf; // save bytes until more data arrives
         buf = null; // force do-while loop to exit
 
-        loggerRecvBytes.finest("wait for more, leftover: {leftover.length} bytes");
+        loggerRecvBytes
+            .finest("wait for more, leftover: ${leftover.length} bytes");
       }
     } while (buf != null);
 
@@ -126,17 +132,19 @@ StreamTransformer<Uint8List, LDAPMessage> _createLdapTransformer() {
     if (error is TlsException) {
       TlsException e = error;
       if (e.osError == null &&
-          e.message == "OSStatus = -9805: connection closed gracefully error -9805") {
+          e.message ==
+              "OSStatus = -9805: connection closed gracefully error -9805") {
         // Connection closed gracefully: ignore this, usually due to application
         // deliberately closing the connection
         if (leftover == null) {
           // Clean close
-          loggerRecvBytes.info("Connection closed gracefully: no leftover bytes to parse");
+          loggerRecvBytes
+              .info("Connection closed gracefully: no leftover bytes to parse");
           return;
         }
 
-        loggerRecvBytes.info("Connection closed gracefully: has leftover bytes to parse");
-
+        loggerRecvBytes
+            .info("Connection closed gracefully: has leftover bytes to parse");
       }
     }
     loggerRecvBytes.severe("LDAP stream transformer: error=${error}");